refactor: remaining util files
