+++ src/util/date.ts
@@ -1,11 +1,7 @@
 import { BigInt } from "@graphprotocol/graph-ts";
 import { secondsInDay } from "../config";
 
-export function toBigInt(integer: i32): BigInt {
-  return BigInt.fromI32(integer);
-}
-
 export function timestampToDate(timestamp: BigInt): BigInt {
-  let daysSinceEpochStart: BigInt = timestamp / BigInt.fromI32(secondsInDay);
-  return daysSinceEpochStart * BigInt.fromI32(secondsInDay);
+  let daysSinceEpochStart: BigInt = timestamp.div(BigInt.fromI32(secondsInDay))
+  return daysSinceEpochStart.times(BigInt.fromI32(secondsInDay))
 }
